Revert fix for multi-clicks in Tracker
Reverts part of 8d0464e.
That fix resulted in menus collapsing as soon as the mouse is moved.

Reverting this part fixes this, at the cost of double-clicks not working
in Tracker. Until fixed, this is acceptable, because you can just hit
ENTER to open a selected file. And ALT+CTRL+H hide a window, same as a
double-click on a window tab.

Fixes #3
Reopens #1